+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 0.1.17
 
-  * 
+  * Add comparison to Evening
 
 ## Version 0.1.16